requirements.txt
fastapi==0.111.0
google-cloud-bigquery==3.22.0
gunicorn==22.0.0
uvicorn==0.29.0
pytest==8.2.0
